Upstream version 2.8.0

# gpg: Signature made Thu Jul 25 11:20:08 2013 JST using RSA key ID 31496D58
# gpg: Good signature from "Thomas Moulard <thomas.moulard@gmail.com>"
# gpg:                 aka "[jpeg image of size 1485]"

+++ src/device/display/windows/vpWin32Window.cpp
@@ -1,6 +1,6 @@
 /****************************************************************************
  *
- * $Id: vpWin32Window.cpp 4073 2013-01-22 09:58:17Z fspindle $
+ * $Id: vpWin32Window.cpp 4304 2013-07-04 14:29:24Z fspindle $
  *
  * This file is part of the ViSP software.
  * Copyright (C) 2005 - 2013 by INRIA. All rights reserved.
@@ -185,7 +185,7 @@ LRESULT CALLBACK WndProc(HWND hWnd, UINT message, WPARAM wParam, LPARAM lParam)
     case WM_KEYDOWN: 
       //case WM_KEYUP: 
       {
-	GetKeyNameText(lParam, window->lpString, 10); // 10 is the size of lpString
+    GetKeyNameText((LONG)lParam, window->lpString, 10); // 10 is the size of lpString
 	//window->key = MapVirtualKey(wParam, MAPVK_VK_TO_CHAR);
 	vpReleaseSemaphore(window->semaKey,1,NULL);
 	break;
