+++ src/tools/histogram/vpHistogram.cpp
@@ -1,6 +1,6 @@
 /****************************************************************************
  *
- * $Id: vpHistogram.cpp 4056 2013-01-05 13:04:42Z fspindle $
+ * $Id: vpHistogram.cpp 4303 2013-07-04 14:14:00Z fspindle $
  *
  * This file is part of the ViSP software.
  * Copyright (C) 2005 - 2013 by INRIA. All rights reserved.
@@ -857,7 +857,7 @@ unsigned vpHistogram::sort(std::list<vpHistogramPeak> & peaks)
 
   peaks.sort(compare_vpHistogramPeak);
 
-  return peaks.size();
+  return (unsigned int)peaks.size();
 }
 
 /*!
